requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
frozenlist==1.8.0
greenlet==3.3.1
idna==3.11
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.7.1
playwright==1.57.0
propcache==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyee==13.0.0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.22.0
